pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sqlkit"
version = "0.0.1"
description = "Object-oriented SQL query construction with multi-database dialect support"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [
    { name = "SQLKit Team" },
]
keywords = [
    "sql",
    "query",
    "database",
    "sqlalchemy",
    "orm",
    "mysql",
    "postgresql",
    "sqlite",
    "redshift",
    "athena",
    "oracle"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "sqlalchemy>=2.0.0",
    "jinja2>=3.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.17.0",
    "pre-commit>=4.0.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.0.0",
    "ruff>=0.12.4",
    "types-PyYAML>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/sqlkit-team/sqlkit"
Documentation = "https://sqlkit.readthedocs.io"
Repository = "https://github.com/sqlkit-team/sqlkit.git"
Issues = "https://github.com/sqlkit-team/sqlkit/issues"

[tool.black]
line-length = 79
target-version = ['py313']

[tool.ruff]
line-length = 79
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming
    "UP", # pyupgrade
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "tests/",
]
